+++ translations/bt_nb.ts
@@ -7011,6 +7011,14 @@ Sluttvolumet i primærgjæringskaret er %1.</translation>
         <source>Amount Type</source>
         <translation type="unfinished">Enhetstype</translation>
     </message>
+    <message>
+        <source>ID in database</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>The ID of this fermentable in the database.  Sometimes useful for debugging.</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>fermentationEditor</name>
@@ -7079,11 +7087,11 @@ Sluttvolumet i primærgjæringskaret er %1.</translation>
     </message>
     <message>
         <source>Temp. lag time</source>
-        <translation type="unfinished">Temp. forsinkelsestid</translation>
+        <translation type="obsolete">Temp. forsinkelsestid</translation>
     </message>
     <message>
         <source>Lag time</source>
-        <translation type="unfinished">Forsinkelsestid</translation>
+        <translation type="obsolete">Forsinkelsestid</translation>
     </message>
     <message>
         <source>End temp.</source>