+++ translations/bt_ru.ts
@@ -7074,6 +7074,14 @@ The final volume in the primary is %1.</source>
         <source>Amount Type</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>ID in database</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>The ID of this fermentable in the database.  Sometimes useful for debugging.</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>fermentationEditor</name>
@@ -7142,11 +7150,11 @@ The final volume in the primary is %1.</source>
     </message>
     <message>
         <source>Temp. lag time</source>
-        <translation type="unfinished">Температура лаг-паузы</translation>
+        <translation type="obsolete">Температура лаг-паузы</translation>
     </message>
     <message>
         <source>Lag time</source>
-        <translation type="unfinished">Лаг-пауза</translation>
+        <translation type="obsolete">Лаг-пауза</translation>
     </message>
     <message>
         <source>End temp.</source>