+++ translations/bt_zh.ts
@@ -6844,6 +6844,14 @@ The final volume in the primary is %1.</source>
         <source>Amount Type</source>
         <translation type="unfinished">金额类型MiscTableModel |</translation>
     </message>
+    <message>
+        <source>ID in database</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>The ID of this fermentable in the database.  Sometimes useful for debugging.</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>fermentationEditor</name>
@@ -6912,11 +6920,11 @@ The final volume in the primary is %1.</source>
     </message>
     <message>
         <source>Temp. lag time</source>
-        <translation type="unfinished">临时滞后时间Temp. lag time</translation>
+        <translation type="obsolete">临时滞后时间Temp. lag time</translation>
     </message>
     <message>
         <source>Lag time</source>
-        <translation type="unfinished">滞后时间Lag time</translation>
+        <translation type="obsolete">滞后时间Lag time</translation>
     </message>
     <message>
         <source>End temp.</source>