+++ CHANGES.markdown
@@ -21,6 +21,8 @@ Minor bug fixes for the 4.0.3 release (ie bugs in 4.0.3 are fixed in this 4.0.4 
 
 ### Bug Fixes
 * Crash editing Target Boil Size [817](https://github.com/Brewtarget/brewtarget/issues/817)
+* Crash when exporting recipe [821](https://github.com/Brewtarget/brewtarget/issues/821)
+* Inventory not displaying (further fixes) [814](https://github.com/Brewtarget/brewtarget/issues/814)
 
 ### Release Timestamp
 Wed, 28 Aug 2024 04:00:04 +0100